backend/requirements.txt
fastapi
uvicorn[standard]
rdkit-pypi
numpy
